agentic-rag-backend/python_rag/requirements.txt
openai
google-generativeai
python-dotenv
pymongo
sentence-transformers
tavily-python
